Extract loading equipment from inventory changer config to loadEquipmentFromJson() function

+++ ftp_cheat/Hacks/Aimbot.cpp
@@ -2,6 +2,7 @@
 #include <array>
 #include <initializer_list>
 #include <memory>
+#include <cmath>
 
 #include "Aimbot.h"
 #include "../Config.h"
@@ -198,7 +199,7 @@ static bool canScan(Entity* entity, const Vector& destination, const WeaponInfo*
             break;
 
         if (trace.entity == entity && trace.hitgroup > HitGroup::Generic && trace.hitgroup <= HitGroup::RightLeg) {
-            damage = HitGroup::getDamageMultiplier(trace.hitgroup) * damage * powf(weaponData->rangeModifier, trace.fraction * weaponData->range / 500.0f);
+            damage = HitGroup::getDamageMultiplier(trace.hitgroup) * damage * std::pow(weaponData->rangeModifier, trace.fraction * weaponData->range / 500.0f);
 
             if (float armorRatio{ weaponData->armorRatio / 2.0f }; HitGroup::isArmored(trace.hitgroup, trace.entity->hasHelmet()))
                 damage -= (trace.entity->armor() < damage * armorRatio / 2.0f ? trace.entity->armor() * 4.0f : damage) * (1.0f - armorRatio);
